+++ viable-veganism/app/components/carousel.js
@@ -2,7 +2,7 @@ import React, { useState, useRef } from 'react';
 import Slider from 'react-slick';
 import 'slick-carousel/slick/slick.css';
 import 'slick-carousel/slick/slick-theme.css';
-import './Carousel.css';
+import './css/carousel.css';
 
 export default function Carousel() {
   const slides = [
