+++ viable-veganism/app/components/dropdown.js
@@ -1,6 +1,6 @@
 import React from "react";
 import {Accordion, AccordionItem} from "@nextui-org/react";
-import './dropdown.css';
+import './css/dropdown.css';
 
 export default function DropDownTabs() {
   const accordionStyle = {
