+++ viable-veganism/package.json
@@ -9,7 +9,9 @@
     "lint": "next lint"
   },
   "dependencies": {
+    "@emailjs/browser": "^4.3.3",
     "@nextui-org/react": "^2.2.10",
+    "emailjs-com": "^3.2.0",
     "next": "14.1.0",
     "react": "^18",
     "react-dom": "^18",
